poc-8-backlog-to-goals/pyproject.toml
[project]
name = "poc-8-backlog-to-goals"
version = "0.1.0"
description = "Proof of Concept 8: Parse backlog markdown into a structured directory of goal description files, using LLM for title sanitization."
authors = [
    {name = "AI Agent / Thomas Silloway"},
]
dependencies = [
    "pydantic>=2.0,<3.0",
    "pydantic-ai>=2.0.0", # Check for the latest appropriate version
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9" # LlmPromptService uses async/await, good to specify modern Python.

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]  # Add src to pythonpath so tests can import from src
asyncio_mode = "auto" # Use 'auto' mode for pytest-asyncio
testpaths = [
    "tests",
]
# Add other pytest configurations if needed, like markers, etc.
# log_cli = true # Example: to show print statements and logs during tests
# log_cli_level = "INFO"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# Optional: for linters or formatters if you decide to add them
# [tool.ruff]
# line-length = 88
# select = ["E", "F", "W", "I", "UP", "PL", "T20"] # Example selection
